fixed running make problem caused by docs

+++ trunk/source/2.0/bin/docbookPrep.sh
@@ -5,9 +5,9 @@ XML_VERSION='4.4'
 XML_FILE="docbook-xml-${XML_VERSION}"
 XML_URL="http://www.docbook.org/xml/${XML_VERSION}/${XML_FILE}.zip"
 
-XSL_VERSION='1.71.0'
+XSL_VERSION='1.72.0'
 XSL_FILE="docbook-xsl-${XSL_VERSION}"
-XSL_URL="http://prdownloads.sourceforge.net/docbook/${XSL_FILE}.tar.bz2?download"
+XSL_URL="http://downloads.sourceforge.net/docbook/${XSL_FILE}.tar.bz2"
 
 #-----------------------------------------------------------------------------
 # do no edit below here
@@ -16,7 +16,7 @@ XSL_URL="http://prdownloads.sourceforge.net/docbook/${XSL_FILE}.tar.bz2?download
 PATH="${PATH}:${MY_DIR}"
 PWD=${PWD:-`pwd`}
 
-MY_BASE=`basename $0`
+MY_BASE=`basename "$0"`
 MY_DIR=`dirname "$0"`
 
 # load shlib
@@ -61,27 +61,37 @@ echo 'Docbook XML'
 echo '  downloading'
 cd ${DL_DIR}
 get_url "${XML_URL}"
-echo '  extracting'
-xml_dir="${DOCBOOK_DIR}/docbook-xml/${XML_VERSION}"
-rm -fr "${xml_dir}"
-mkdir -p "${xml_dir}"
-cd "${xml_dir}"
-unzip -oq "${DL_DIR}/${XML_FILE}.zip"
-cd ..
-rm -f current
-ln -s "${XML_VERSION}" current
+if [ -f "${DL_DIR}/${XML_FILE}.zip" ]; then
+  echo '  extracting'
+  xml_dir="${DOCBOOK_DIR}/docbook-xml/${XML_VERSION}"
+  rm -fr "${xml_dir}"
+  mkdir -p "${xml_dir}"
+  cd "${xml_dir}"
+  unzip -oq "${DL_DIR}/${XML_FILE}.zip"
+  cd ..
+  rm -f current
+  ln -s "${XML_VERSION}" current
+else
+  echo "error: unable to extract (${XML_FILE}.zip)" >&2
+  exit 1
+fi
 
 # get the docbook xslt files
 echo 'Docbook XSLT'
 echo '  downloading'
 cd ${DL_DIR}
 get_url "${XSL_URL}"
-echo '  extracting'
-xsl_dir="${DOCBOOK_DIR}/docbook-xsl"
-mkdir -p "${xsl_dir}"
-cd "${xsl_dir}"
-rm -fr ${XSL_VERSION}
-bzip2 -dc ${DL_DIR}/${XSL_FILE}.tar.bz2 |tar xf -
-mv ${XSL_FILE} ${XSL_VERSION}
-rm -f current
-ln -s "${XSL_VERSION}" current
+if [ -f "${DL_DIR}/${XSL_FILE}.tar.bz2" ]; then
+  echo '  extracting'
+  xsl_dir="${DOCBOOK_DIR}/docbook-xsl"
+  mkdir -p "${xsl_dir}"
+  cd "${xsl_dir}"
+  rm -fr ${XSL_VERSION}
+  bzip2 -dc "${DL_DIR}/${XSL_FILE}.tar.bz2" |tar xf -
+  mv ${XSL_FILE} ${XSL_VERSION}
+  rm -f current
+  ln -s "${XSL_VERSION}" current
+else
+  echo "error: unable to extract (${XSL_FILE}.tar.bz2)" >&2
+  exit 1
+fi
